# portfolio_api_swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Investment Portfolio Tracker API",
    "version": "1.0.0",
    "description": "REST API for managing investment portfolios, holdings, and transactions."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
  "paths": {
    "/api/auth/signup": {
      "post": {
        "summary": "Create a new user account",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "user@example.com" },
                  "password": { "type": "string", "minLength": 6, "example": "Password123!" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created" },
          "400": { "description": "Bad Request" },
          "409": { "description": "Conflict" }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "summary": "Authenticate a user and return a JWT",
        "tags": ["Auth"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email": { "type": "string", "format": "email", "example": "owner@example.com" },
                  "password": { "type": "string", "example": "Password123!" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "OK" },
          "401": { "description": "Unauthorized" }
        }
      }
    },
    "/api/portfolios": {
      "get": {
        "summary": "Get all portfolios",
        "tags": ["Portfolios"],
        "security": [{ "BearerAuth": [] }],
        "responses": { "200": { "description": "OK" } }
      },
      "post": {
        "summary": "Create a portfolio",
        "tags": ["Portfolios"],
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "example": "My Portfolio" },
                  "description": { "type": "string", "example": "Test" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created" } }
      }
    },
    "/api/holdings": {
      "get": {
        "summary": "Get holdings",
        "tags": ["Holdings"],
        "security": [{ "BearerAuth": [] }],
        "parameters": [{ "name": "portfolioId", "in": "query", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "OK" } }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}